Re-renamespace renderers
