+++ src/Services/ResultPrinter/ResultPrinter.php
@@ -17,9 +17,9 @@ use webignition\BasilRunner\Model\TestOutput\Test as TestOutput;
 use webignition\BasilRunner\Services\ProjectRootPathProvider;
 use webignition\BasilRunner\Services\ResultPrinter\FailedAssertion\SummaryLineFactory;
 use webignition\BasilRunner\Services\ResultPrinter\FailedAssertion\SummaryHandler;
-use webignition\BasilRunner\Services\TestOutputRenderer\StatementLineRenderer;
-use webignition\BasilRunner\Services\TestOutputRenderer\StepRenderer;
-use webignition\BasilRunner\Services\TestOutputRenderer\TestRenderer;
+use webignition\BasilRunner\Services\ResultPrinter\Renderer\StatementLineRenderer;
+use webignition\BasilRunner\Services\ResultPrinter\Renderer\StepRenderer;
+use webignition\BasilRunner\Services\ResultPrinter\Renderer\TestRenderer;
 
 class ResultPrinter extends Printer implements TestListener
 {
